+++ plugins/axios.ts
@@ -4,6 +4,7 @@ export default defineNuxtPlugin({
     name: "axios",
     setup: () => {
         const config = useRuntimeConfig();
+        const appConfig = useAppConfig();
         const { token } = useAuth();
 
         const axios = axiosOriginal.create({
@@ -11,8 +12,16 @@ export default defineNuxtPlugin({
         });
 
         axios.interceptors.request.use((config) => {
-            if (config.authorization) {
-                config.headers.Authorization = token.value ? `Bearer ${token.value}` : undefined;
+            if (appConfig.appApi?.extendHeaders) {
+                config.headers = appConfig.appApi.extendHeaders(config.headers);
+            }
+
+            if (config.authorization && token.value) {
+                if (appConfig.appApi?.customAuthorizationHeader) {
+                    appConfig.appApi.customAuthorizationHeader(token.value);
+                } else {
+                    config.headers.Authorization = `Bearer ${token.value}`;
+                }
             }
 
             return config;
@@ -23,7 +32,7 @@ export default defineNuxtPlugin({
                 return response;
             },
             async function (error: ApiError) {
-                if (process.env.NODE_ENV === 'development') {
+                if (process.env.NODE_ENV === "development") {
                     console.log(
                         `[ AXIOS_ERROR ] - [ ${error.response?.config.url} ] - [${error.status}] \n`,
                         error.response?.data